+++ spec/kettle/family/cli_spec.rb
@@ -611,7 +611,6 @@ RSpec.describe Kettle::Family::CLI do
         "MOLINILLO_DEBUG",
         "#{File.basename(@tmpdir).gsub(/[^A-Za-z0-9]+/, "_").upcase}_DEV=#{@tmpdir}",
         "BUNDLE_DISABLE_CHECKSUM_VALIDATION=true",
-        "KETTLE_JEM_THREAD_WORKERS=#{[1, Etc.nprocessors - 1].max}",
         "KETTLE_JEM_GIT_LOCK=#{File.join(@tmpdir, ".git", "kettle-family-template-commit.lock")}",
         "KETTLE_JEM_GIT_COMMIT_LOCK=#{File.join(@tmpdir, ".git", "kettle-family-template-commit.lock")}",
         "K_JEM_TEMPLATING=true",
@@ -627,6 +626,7 @@ RSpec.describe Kettle::Family::CLI do
         "BUNDLE_SILENCE_DEPRECATIONS=true",
         "BUNDLE_SILENCE_ROOT_WARNING=true",
         "BUNDLE_SUPPRESS_INSTALL_USING_MESSAGES=true",
+        "KETTLE_FAMILY_WAVE_JOBS=1",
         "sh",
         "-lc",
         "kettle-jem install --quiet --events"